MOV: Helpers Functions to www_som
